Comments on MongoDBWrapper and new functions that are used in our new CalculateBaselines() func.
In the XMLParser I removed the overwriting of the szenarioName with the entry from the .dtp file because for some reason Dexterit sometimes added a 'u' before the actual szenarioname in the .dtp file. We don't want this and the szenarioName from the c3d file never has the 'u' so now we just take that one.

Included some code in the ManipAnalysisFunctions to automatically parse the Baseline. Not fully finished yet!